# core/api/gen/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'g': // Prefix: "get"

				if l := len("get"); len(elem) >= l && elem[0:l] == "get" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'A': // Prefix: "Aggregation"

					if l := len("Aggregation"); len(elem) >= l && elem[0:l] == "Aggregation" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'P': // Prefix: "Proof"

						if l := len("Proof"); len(elem) >= l && elem[0:l] == "Proof" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetAggregationProofGetRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}

					case 'S': // Prefix: "Status"

						if l := len("Status"); len(elem) >= l && elem[0:l] == "Status" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetAggregationStatusGetRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}

					}

				case 'C': // Prefix: "CurrentEpoch"

					if l := len("CurrentEpoch"); len(elem) >= l && elem[0:l] == "CurrentEpoch" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetCurrentEpochGetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}

				case 'S': // Prefix: "S"

					if l := len("S"); len(elem) >= l && elem[0:l] == "S" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "ignature"

						if l := len("ignature"); len(elem) >= l && elem[0:l] == "ignature" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'R': // Prefix: "Request"

							if l := len("Request"); len(elem) >= l && elem[0:l] == "Request" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "GET":
									s.handleGetSignatureRequestGetRequest([0]string{}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, "GET")
								}

								return
							}

						case 's': // Prefix: "s"

							if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "GET":
									s.handleGetSignaturesGetRequest([0]string{}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, "GET")
								}

								return
							}

						}

					case 'u': // Prefix: "uggestedEpoch"

						if l := len("uggestedEpoch"); len(elem) >= l && elem[0:l] == "uggestedEpoch" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetSuggestedEpochGetRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}

					}

				case 'V': // Prefix: "ValidatorSet"

					if l := len("ValidatorSet"); len(elem) >= l && elem[0:l] == "ValidatorSet" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetValidatorSetGetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}

				}

			case 's': // Prefix: "signMessage"

				if l := len("signMessage"); len(elem) >= l && elem[0:l] == "signMessage" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleSignMessagePostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "POST")
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name        string
	summary     string
	operationID string
	pathPattern string
	count       int
	args        [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'g': // Prefix: "get"

				if l := len("get"); len(elem) >= l && elem[0:l] == "get" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'A': // Prefix: "Aggregation"

					if l := len("Aggregation"); len(elem) >= l && elem[0:l] == "Aggregation" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'P': // Prefix: "Proof"

						if l := len("Proof"); len(elem) >= l && elem[0:l] == "Proof" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetAggregationProofGetOperation
								r.summary = "Get aggregation proof"
								r.operationID = ""
								r.pathPattern = "/getAggregationProof"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'S': // Prefix: "Status"

						if l := len("Status"); len(elem) >= l && elem[0:l] == "Status" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetAggregationStatusGetOperation
								r.summary = "Get aggregation status, can be sent only to aggregator nodes"
								r.operationID = ""
								r.pathPattern = "/getAggregationStatus"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'C': // Prefix: "CurrentEpoch"

					if l := len("CurrentEpoch"); len(elem) >= l && elem[0:l] == "CurrentEpoch" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetCurrentEpochGetOperation
							r.summary = "Get current epoch"
							r.operationID = ""
							r.pathPattern = "/getCurrentEpoch"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'S': // Prefix: "S"

					if l := len("S"); len(elem) >= l && elem[0:l] == "S" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "ignature"

						if l := len("ignature"); len(elem) >= l && elem[0:l] == "ignature" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'R': // Prefix: "Request"

							if l := len("Request"); len(elem) >= l && elem[0:l] == "Request" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "GET":
									r.name = GetSignatureRequestGetOperation
									r.summary = "Get signature request by request hash"
									r.operationID = ""
									r.pathPattern = "/getSignatureRequest"
									r.args = args
									r.count = 0
									return r, true
								default:
									return
								}
							}

						case 's': // Prefix: "s"

							if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "GET":
									r.name = GetSignaturesGetOperation
									r.summary = "Get signature by request hash"
									r.operationID = ""
									r.pathPattern = "/getSignatures"
									r.args = args
									r.count = 0
									return r, true
								default:
									return
								}
							}

						}

					case 'u': // Prefix: "uggestedEpoch"

						if l := len("uggestedEpoch"); len(elem) >= l && elem[0:l] == "uggestedEpoch" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetSuggestedEpochGetOperation
								r.summary = "Get suggested epoch to request sign"
								r.operationID = ""
								r.pathPattern = "/getSuggestedEpoch"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'V': // Prefix: "ValidatorSet"

					if l := len("ValidatorSet"); len(elem) >= l && elem[0:l] == "ValidatorSet" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetValidatorSetGetOperation
							r.summary = "Get current validator set"
							r.operationID = ""
							r.pathPattern = "/getValidatorSet"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 's': // Prefix: "signMessage"

				if l := len("signMessage"); len(elem) >= l && elem[0:l] == "signMessage" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = SignMessagePostOperation
						r.summary = "Sign a message"
						r.operationID = ""
						r.pathPattern = "/signMessage"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}